third_party/WebKit/LayoutTests/external/wpt/encoding/big5-encoder-expected.txt

This is a testharness.js-based test.
PASS big5 encoder: very basic
FAIL big5 encoder: Highest-pointer BMP character excluded from encoder assert_equals: expected "X&%2340614;X" but got "X%26%2340614%3BX"
FAIL big5 encoder: Highest-pointer character excluded from encoder assert_equals: expected "X&%23156267;X" but got "X%26%23156267%3BX"
PASS big5 encoder: Lowest-pointer character included in encoder
PASS big5 encoder: Euro; the highest-pointer character before a range of 30 unmapped pointers
PASS big5 encoder: The lowest-pointer character after the range of 30 unmapped pointers
PASS big5 encoder: The highest-pointer character before a range of 41 unmapped pointers
PASS big5 encoder: The lowest-pointer character after the range of 41 unmapped pointers
PASS big5 encoder: The last character in the index
FAIL big5 encoder: The canonical BMP test character that is not in the index assert_equals: expected "X&%239731;X" but got "X%26%239731%3BX"
FAIL big5 encoder: The canonical astral test character that is not in the index assert_equals: expected "X&%23128169;X" but got "X%26%23128169%3BX"
PASS big5 encoder: A Plane 2 character whose low 16 bits match a BMP character that has a lower pointer
PASS big5 encoder: A duplicate-mapped code point that prefers the highest pointer in the encoder
Harness: the test ran to completion.
